Renamed pages components to include to 'Page' suffix
