+++ frontend/src/App.js
@@ -2,11 +2,9 @@ import React from "react";
 import { BrowserRouter, Switch, Route } from "react-router-dom";
 
 import AuthState from "./context/auth/AuthState";
-
 import Navbar from "./components/Navbar";
-
-import Home from "./pages/Home";
-import Login from "./pages/Login";
+import HomePage from "./pages/HomePage";
+import LoginPage from "./pages/LoginPage";
 
 const App = () => {
   return (
@@ -14,8 +12,8 @@ const App = () => {
       <BrowserRouter>
         <Navbar />
         <Switch>
-          <Route exact path="/" component={Home} />
-          <Route path="/login" component={Login} />
+          <Route exact path="/" component={HomePage} />
+          <Route path="/login" component={LoginPage} />
         </Switch>
       </BrowserRouter>
     </AuthState>
