+++ frontend/src/pages/HomePage.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { Container } from "@material-ui/core";
 
-const Home = () => {
+const HomePage = () => {
   return <Container>Hello from the Home page</Container>;
 };
 
-export default Home;
+export default HomePage;
